+++ src/routes/routes.ts
@@ -4,3 +4,4 @@ export const HOME = `/${ROOT}/`;
 export const CREATE_TAG = `/${ROOT}/tag/`;
 export const EDIT_TODO = `/${ROOT}/:id/`;
 export const EDIT_TAGS = `/${ROOT}/tags/`;
+export const EDIT_TAG = `/${ROOT}/tags/:id`;